+++ vivarium/core/emitter.py
@@ -418,12 +418,14 @@ def get_history_data_db(
     if query:
         projection = {f"data.{'.'.join(field)}": 1 for field in query}
         projection['data.time'] = 1
+        projection['assembly_id'] = 1
 
     cursor = history_collection.find(experiment_query, projection)
     raw_data = []
     for document in cursor:
-        if document['data'].get('time'):
-            raw_data.append(document)
+        assert document.get('assembly_id'), \
+            "all database documents require an assembly_id"
+        raw_data.append(document)
 
     # re-assemble data
     assembly = assemble_data(raw_data)